misc(app): Deduplicate button and product cards components
