+++ src/screens/CartScreen.tsx
@@ -1,22 +1,14 @@
 import * as React from 'react';
-import {
-  Button,
-  View,
-  StyleSheet,
-  Text,
-  ActivityIndicator,
-  FlatList,
-} from 'react-native';
+import {Button, View, StyleSheet, Text, FlatList} from 'react-native';
 import {useSelector, useDispatch} from 'react-redux';
 import * as Sentry from '@sentry/react-native';
 import Toast from 'react-native-toast-message';
-import LinearGradient from 'react-native-linear-gradient';
-import {TouchableOpacity} from 'react-native-gesture-handler';
-import {RootState, AppDispatch} from '../reduxApp';
-import {selectImage} from './EmpowerPlant';
+import {RootState} from '../reduxApp';
 import {BACKEND_URL} from '../config';
 import {RootStackParamList} from '../navigation';
 import {StackScreenProps} from '@react-navigation/stack';
+import {ProfiledStyledCartProductCard} from '../components/StyledCartProductCard';
+import { Product } from '../types/Product';
 
 interface CartData {
   name: string;
@@ -148,7 +140,7 @@ const CartScreen = ({
           data={Object.values(cartData)}
           renderItem={({item}) => {
             return (
-              <ProfiledCartItem
+              <ProfiledStyledCartProductCard
                 appDispatch={dispatch}
                 quantity={item.quantity}
                 title={item.title}
@@ -167,94 +159,14 @@ const CartScreen = ({
 };
 export default Sentry.withProfiler(CartScreen);
 
-export const GradientBtn = (props: {
-  colors: Array<string>;
-  style: any;
-  name: string;
-  onPress: any;
-  progressState: boolean;
-  buttonText: any;
-}): React.ReactElement => {
-  //dependencies for this may also need to be added for android
-  //linear gradient colors Array<string> requires 2 colors
-  return (
-    <TouchableOpacity onPress={props.onPress}>
-      <LinearGradient style={props.style} colors={props.colors}>
-        {props.progressState ? (
-          <ActivityIndicator size="small" color="#404091" />
-        ) : (
-          <Text style={props.buttonText}>{props.name}</Text>
-        )}
-      </LinearGradient>
-    </TouchableOpacity>
-  );
-};
-
-const CartItem = (props: {
-  imgcropped: string;
-  id: number;
-  quantity: number;
-  price: number;
-  appDispatch: AppDispatch;
-  title: string;
-}): React.ReactElement => {
-  const deleteItem = (id: string) => {
-    props.appDispatch({type: 'DELETE_FROM_CART', payload: id});
-  };
-
-  return (
-    <View style={styles.statisticContainer}>
-      <View>{selectImage(props.imgcropped)}</View>
-      <View>
-        <Text style={styles.itemTitle}>
-          {props.title.charAt(0).toUpperCase() + props.title.slice(1)}
-        </Text>
-
-        {/* TODO <Text style={styles.sku}>{'sku: ' + props.sku}</Text> */}
-        <Text style={styles.itemPrice}>
-          {'$' + props.price + ` (${props.quantity})`}
-        </Text>
-        <GradientBtn
-          buttonText={styles.buttonText}
-          colors={['#ebebeb', '#b3b3b3']}
-          style={styles.deleteBtn}
-          name={'Delete'}
-          progressState={false}
-          onPress={() => deleteItem(props.id.toString())}
-        />
-      </View>
-    </View>
-  );
-};
-
-const ProfiledCartItem = Sentry.withProfiler(CartItem);
-
 const styles = StyleSheet.create({
   contentContainer: {
     paddingBottom: 165,
   },
-  sku: {
-    fontSize: 16,
-    color: '#002626',
-    marginBottom: 10,
-  },
   screen: {
     flex: 1,
     flexDirection: 'column',
-
     backgroundColor: '#ffff',
-  },
-  flavorContainer: {
-    width: '100%',
-    backgroundColor: '#f5f5f5',
-    padding: 10,
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'flex-end',
-  },
-  logo: {
-    width: 20,
-    height: 20,
   },
   titleContainer: {
     justifyContent: 'center',
@@ -262,84 +174,10 @@ const styles = StyleSheet.create({
     borderBottomWidth: 1,
     borderColor: '#dbdbdb',
   },
-  title: {
-    fontSize: 24,
-    fontWeight: '700',
-  },
-  tinyImage: {
-    width: 100,
-    height: 150,
-  },
-  card: {
-    width: '100%',
-    height: 240,
-    padding: 12,
-    borderWidth: 1,
-    borderColor: '#79628C',
-    borderRadius: 6,
-
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  statisticContainer: {
-    width: '100%',
-    height: 240,
-    padding: 12,
-    borderBottomWidth: 1,
-
-    borderColor: '#dbdbdb',
-    flex: 1,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-around',
-    marginVertical: 10,
-  },
-  statisticTitle: {
-    fontSize: 16,
-    fontWeight: '500',
-  },
-  statisticCount: {
-    fontSize: 16,
-    fontWeight: '700',
-  },
-  deleteBtn: {
-    height: 40,
-    width: 100,
-    marginTop: 10,
-    paddingLeft: 10,
-    paddingRight: 10,
-    borderRadius: 2,
-    borderWidth: 1,
-    borderColor: '#808080',
-    flexDirection: 'column',
-    justifyContent: 'center',
-  },
-  linearGradient: {
-    height: 50,
-
-    paddingLeft: 20,
-    paddingRight: 20,
-    borderRadius: 2,
-    borderWidth: 1,
-    borderColor: '#8D6E63',
-    flexDirection: 'column',
-    justifyContent: 'center',
-  },
   buttonText: {
     textAlign: 'center',
     fontSize: 16,
     color: 'white',
-  },
-  itemTitle: {
-    marginBottom: 5,
-    fontSize: 17,
-    fontWeight: '500',
-    color: '#002626',
-  },
-  itemPrice: {
-    fontSize: 22,
-    fontWeight: '400',
-    color: '#002626',
   },
   subtotalText: {
     marginTop: 20,
