+++ src/screens/CheckoutScreen.tsx
@@ -11,8 +11,8 @@ import {useDispatch, useSelector} from 'react-redux';
 import * as Sentry from '@sentry/react-native';
 import Toast from 'react-native-toast-message';
 import {RootState} from '../reduxApp';
-import {GradientBtn} from './CartScreen';
 import {BACKEND_URL} from '../config';
+import {StyledButton} from '../components/StyledButton';
 
 export type UIToast = typeof Toast;
 
@@ -136,13 +136,10 @@ const CheckoutScreen = () => {
           </Text>
         </View>
         <View>
-          <GradientBtn
-            buttonText={styles.buttonText}
-            colors={['#002626', '#001414']}
-            style={styles.linearGradient}
+          <StyledButton
             onPress={() => performCheckoutOnServer()}
-            progressState={orderStatusUI}
-            name={'Place your order'}
+            isLoading={orderStatusUI}
+            title={'Place your order'}
           />
         </View>
       </View>
