+++ src/types/Product.ts
@@ -1,0 +1,11 @@
+
+export type Product = {
+  sku: string;
+  name: string;
+  image: string;
+  id: number;
+  type: string;
+  price: number;
+  title: string;
+  imgcropped: string;
+};
